configs: Accept a "type" argument in "output" blocks
This uses the same type constraint expression syntax as we use in the
argument of the same name in variable blocks. It doesn't do anything yet,
but in subsequent commits it should become similar in purpose to type
constraints on variables, allowing module authors to specify exactly how
their result types are guaranteed to be constrained and have Terraform
automatically convert the result to match the given constraint.